Split out testSample function to separate doc. #2
